refactor to wrap around resolver, not mapper
This commit addresses author's comments on PR #158 to refactor wrapping
and split prefix logic into a separate file.
